Fix 13 safety audit findings: fault handling, data races, overflow, dead code
Safety fixes applied across firmware and supporting infrastructure:

- stm32g4xx_it.c: fault handlers now call NVIC_SystemReset() instead of looping
  forever; TIM5 PWM zeroed at register level (HAL unsafe in fault context)
- main.c: ms5837 init failure sets pressure_valid=false; ADC4 VREFINT+BAT reads
  made atomic under __disable_irq(); removed unused globals (something,
  vbias_effective_mv, vbias_counts_for_adc2)
- pressure_sensor.c: int64_t intermediate prevents overflow at ~30-bar depth limit
- button.c: button_get_event() read-clear made atomic under __disable_irq()
- calibration.c: bounds check on SensorID before array write
- ext_flash_storage.c: 512-byte local array in ext_cal_iterate() made static
- ms5837.c: PROM CRC4 validation per datasheet §2.4; returns MS5837_ERROR_CRC
  on mismatch
- sensor.c/sensor.h: corrected MAX_STABILITY_WAIT comment; removed duplicate
  include guard
- screen_sensor_data.c: removed redundant local extern declaration
- IWDG enabled via CubeMX (.ioc, hal_conf.h, .project, .mxproject)

Also adds:
- .pipeline/: static analysis pipeline config, findings tracking, safety audit
  report with full resolution log (13 fixed, 2 pre-existing/N/A, 1 deferred)
- tests/: unit test suite covering calibration, drift tracker, flash, ms5837,
  pressure sensor, settings, warning modules
- README.md: Safety Audit section with link to .pipeline/safety_audit.md

Build: clean, text=214500 data=148 bss=57396 (128KB RAM, ~73KB headroom).

+++ STM32CubeIDE/Application/User/Core/pressure_sensor.c
@@ -103,7 +103,7 @@ int32_t pressure_sensor_calculate_depth_mm(int32_t current_pressure_mbar, int32_
 
     /* Apply seawater depth formula with integer arithmetic */
     /* depth_mm = (delta_pressure_mbar × 995) / 100 */
-    int32_t depth_mm = (delta_pressure_mbar * DEPTH_COEFFICIENT_NUMERATOR) / DEPTH_COEFFICIENT_DENOMINATOR;
+    int32_t depth_mm = (int32_t)(((int64_t)delta_pressure_mbar * DEPTH_COEFFICIENT_NUMERATOR) / DEPTH_COEFFICIENT_DENOMINATOR);
 
     return depth_mm;
 }
